Move ChunkSnapshot
