+++ src/world/mod.rs
@@ -87,13 +87,6 @@ pub struct LightUpdate {
 #[derive(PartialEq, Eq, Hash, Clone, Copy)]
 pub struct CPos(pub i32, pub i32);
 
-pub struct ChunkSnapshot {
-    pub position: CPos,
-    pub sections: [Option<SectionSnapshot>; 16],
-    pub biomes: [u8; 16 * 16],
-    pub heightmap: [u8; 16 * 16],
-}
-
 lazy_static! {
     static ref EMPTY_SECTION: SectionSnapshot = SectionSnapshot {
         y: 255, // TODO: Check
